Completely Rendered the Board
With the help of the almighty SAM we are progressing on this conquest to bring less to more. Currently created functions to render the board to the screen and next steps will be to add functionality to the JS and eventually get it on to a server so that we can take ownership over someone else game jk jk don't sue. we are fans.

Co-Authored-By: kemptosa <99417853+kemptosa@users.noreply.github.com>