Pieces on the board!
A majority of this commit is adding coordinates to each spot. Big thanks to Sam for helping with the math end of creating x and y coordinates that transalte to the whole board as it is being rendered.

Also started with rendering the pieces on the board. currently they are hard coded.

The next challenge is going to be creating the logic and grasping the difference between client side rendering and server side to communicate across the web.

The read me has been mainly used for psuedo code  and has mainly been me trying to visualize what is happening and tracking next steps.

Co-Authored-By: kemptosa <99417853+kemptosa@users.noreply.github.com>