Crossing me homie?!?
Through the excellent mind of Samuel Kempton, the deduction was meet that we could use two loops to figure out if we were crossing a piece with our move.

Starting from checking if we are moving horizontally or Vertically with the first set of if statements

then we find the absolute number of moves.

Next using a ternary we determine the direction using negatives or positives.

then a for loop to check all of the spots in-between the two  coordinates.

Also We changed the color of the non walls to a lighter gray as well as the spots.

PROGRESSSSSSSS!!!!!!

Co-Authored-By: kemptosa <99417853+kemptosa@users.noreply.github.com>